refactor: Trasfer -> RentHome, FindRoom -> FindHome 으로 이름 변경
