+++ android/app/src/androidTest/java/com/example/home_rent_app/FindHomeViewModelTest.kt
@@ -2,21 +2,22 @@ package com.example.home_rent_app
 
 import android.content.Intent
 import androidx.test.core.app.ApplicationProvider
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import com.example.home_rent_app.ui.findhome.FindHomeActivity
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 @LargeTest
-class FindRentTest {
+class FindHomeViewModelTest {
+
+    private lateinit var dispatchers: CoroutineDispatcher
 
     @get:Rule
     val activityScenarioRule: ActivityScenarioRule<FindHomeActivity> =
@@ -27,18 +28,13 @@ class FindRentTest {
             )
         )
 
-    @Test
-    fun setAppbarBackButtonTest() {
-        activityScenarioRule.scenario.onActivity {
-            it.onBackPressed()
-            it.finish()
-        }
+    @Before
+    fun setUp() {
+        dispatchers = Dispatchers.Unconfined
     }
 
     @Test
-    fun onViewEditTextTest() {
-        activityScenarioRule.scenario.apply {
-            onView(withId(R.id.tilo_search_room)).check(matches(isDisplayed()))
-        }
+    fun editTextTest() {
+//        Assert.assertThat(viewModel.searchAddress.value, CoreMatchers.`is`("사상구"))
     }
 }
