+++ android/app/src/main/AndroidManifest.xml
@@ -47,7 +47,7 @@
             android:name=".ui.wanthomeresult.WantHomeResultActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.findroom.FindHomeActivity"
+            android:name=".ui.findhome.FindHomeActivity"
             android:exported="false" />
         <activity
             android:name=".ui.SplashScreenActivity"
@@ -60,7 +60,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.transfer.TransferActivity"
+            android:name=".ui.renthome.RentHomeActivity"
             android:exported="false" />
         <activity
             android:name=".ui.wanthome.WantHomeActivity"
