+++ android/app/src/main/java/com/example/home_rent_app/data/datasource/findhome/FindHomeDataSourceImpl.kt
@@ -1,0 +1,31 @@
+package com.example.home_rent_app.data.datasource.findhome
+
+import com.example.home_rent_app.data.api.FindHomeApi
+import com.example.home_rent_app.data.model.BookmarkRequest
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class FindHomeDataSourceImpl @Inject constructor(
+    private val api: FindHomeApi
+) : FindHomeDataSource {
+
+    override fun getSearchResult(
+        page: Int,
+        size: Int,
+        availableOnly: Boolean,
+        sortedBy: String,
+        searchAddress: String
+    ) = flow {
+        emit(api.getSearchResult(page, size, availableOnly, sortedBy, searchAddress))
+    }.flowOn(Dispatchers.IO)
+
+    override suspend fun addBookmark(bookmarkRequest: BookmarkRequest) =
+        api.addBookmark(bookmarkRequest)
+
+    override suspend fun deleteBookmark(bookmarkRequest: BookmarkRequest) =
+        api.deleteBookmark(bookmarkRequest)
+}
