--- android/app/src/main/java/com/example/home_rent_app/data/datasource/transfer/TransferDataSourceImpl.kt
@@ -1,18 +1,0 @@
-package com.example.home_rent_app.data.datasource.transfer
-
-import com.example.home_rent_app.data.api.TransferApi
-import com.example.home_rent_app.data.dto.AddRentHomeRequest
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
-import okhttp3.MultipartBody
-import javax.inject.Inject
-
-class TransferDataSourceImpl @Inject constructor(private val api: TransferApi) : TransferDataSource {
-
-    override fun getImageUrl(list: List<MultipartBody.Part>) = flow {
-        emit(api.getImageUrl(list))
-    }.flowOn(Dispatchers.IO)
-
-    override suspend fun addRentHome(request: AddRentHomeRequest) = api.addRentHome(request)
-}
