+++ android/app/src/main/java/com/example/home_rent_app/data/repository/findhome/FindHomeRepositoryImpl.kt
@@ -1,0 +1,38 @@
+package com.example.home_rent_app.data.repository.findhome
+
+import com.example.home_rent_app.data.datasource.findhome.FindHomeDataSource
+import com.example.home_rent_app.data.dto.toRoomSearchResult
+import com.example.home_rent_app.data.model.BookmarkRequest
+import com.example.home_rent_app.data.model.RoomSearchResult
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class FindHomeRepositoryImpl @Inject constructor(
+    private val dataSource: FindHomeDataSource
+) : FindHomeRepository {
+
+    override fun getSearchResult(
+        page: Int,
+        size: Int,
+        availableOnly: Boolean,
+        sortedBy: String,
+        searchAddress: String
+    ): Flow<RoomSearchResult> {
+        return dataSource.getSearchResult(page, size, availableOnly, sortedBy, searchAddress)
+            .map {
+                it.toRoomSearchResult()
+            }
+    }
+
+    override suspend fun addBookmark(
+        bookmarkRequest: BookmarkRequest
+    ) = dataSource.addBookmark(bookmarkRequest)
+
+    override suspend fun deleteBookmark(
+        bookmarkRequest: BookmarkRequest
+    ) = dataSource.deleteBookmark(bookmarkRequest)
+}
+
