+++ android/app/src/main/java/com/example/home_rent_app/ui/bookmark/BookmarkGiveHomeFragment.kt
@@ -9,17 +9,15 @@ import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.home_rent_app.R
 import com.example.home_rent_app.databinding.FragmentBookmarkGiveHomeBinding
 import com.example.home_rent_app.ui.detail.DetailRentActivity
-import com.example.home_rent_app.ui.findroom.HomeListAdapter
+import com.example.home_rent_app.ui.findhome.HomeListAdapter
 import com.example.home_rent_app.util.ItemIdSession
 import com.example.home_rent_app.util.UserSession
 import com.example.home_rent_app.util.collectStateFlow
-import com.example.home_rent_app.util.logger
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
