+++ android/app/src/main/java/com/example/home_rent_app/ui/home/HomeFragment.kt
@@ -10,10 +10,9 @@ import androidx.fragment.app.Fragment
 import com.example.home_rent_app.R
 import com.example.home_rent_app.databinding.FragmentHomeBinding
 import com.example.home_rent_app.ui.HomeActivity
-import com.example.home_rent_app.ui.findroom.FindHomeActivity
-import com.example.home_rent_app.ui.transfer.TransferActivity
+import com.example.home_rent_app.ui.findhome.FindHomeActivity
+import com.example.home_rent_app.ui.renthome.RentHomeActivity
 import com.example.home_rent_app.ui.wanthomeresult.WantHomeResultActivity
-import kotlinx.coroutines.NonCancellable.start
 
 class HomeFragment : Fragment() {
 
@@ -49,7 +48,7 @@ class HomeFragment : Fragment() {
         }
 
         binding.tvBtnRent.setOnClickListener {
-            val intent = Intent(binding.root.context, TransferActivity::class.java)
+            val intent = Intent(binding.root.context, RentHomeActivity::class.java)
             startActivity(intent)
         }
