--- android/app/src/main/java/com/example/home_rent_app/ui/transfer/step3/AddressSearchFragment.kt
@@ -1,187 +1,0 @@
-package com.example.home_rent_app.ui.transfer.step3
-
-import android.annotation.SuppressLint
-import android.app.Dialog
-import android.net.http.SslError
-import android.os.Bundle
-import android.os.Message
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.webkit.JavascriptInterface
-import android.webkit.JsResult
-import android.webkit.SslErrorHandler
-import android.webkit.WebChromeClient
-import android.webkit.WebResourceRequest
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.navigation.fragment.findNavController
-import com.example.home_rent_app.R
-import com.example.home_rent_app.databinding.FragmentAddressSearchBinding
-import com.example.home_rent_app.ui.viewmodel.TransferViewModel
-import com.example.home_rent_app.util.logger
-import com.example.home_rent_app.util.repeatOnStarted
-import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-
-@AndroidEntryPoint
-class AddressSearchFragment : Fragment() {
-
-    private val binding: FragmentAddressSearchBinding by lazy {
-        FragmentAddressSearchBinding.inflate(layoutInflater)
-    }
-
-    private val viewModel: TransferViewModel by activityViewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.vm = viewModel
-
-        binding.etFindAddress.setOnClickListener {
-            showKakaoAddressWebView()
-        }
-
-        binding.btnNext.setOnClickListener {
-            findNavController().navigate(R.id.action_addressSearchFragment_to_rentHomeDescriptionFragment)
-            viewModel.setNextPage()
-        }
-
-        repeatOnStarted {
-            viewModel.address.collect {
-                viewModel.setAddressPageState()
-                logger("address")
-            }
-        }
-
-        repeatOnStarted {
-            viewModel.addressDetail.collect {
-                viewModel.setAddressPageState()
-                logger("addressDetail")
-            }
-        }
-
-        repeatOnStarted {
-            viewModel.addressPageState.collect {
-                logger("addressPageState")
-                binding.btnNext.isEnabled = it
-            }
-        }
-
-        binding.btnBack.setOnClickListener {
-            findNavController().navigateUp()
-            viewModel.setBackPage()
-        }
-    }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    private fun showKakaoAddressWebView() {
-
-        binding.kakaoAddressWebView.settings.apply {
-            javaScriptEnabled = true
-            javaScriptCanOpenWindowsAutomatically = true
-            setSupportMultipleWindows(true)
-        }
-
-        binding.kakaoAddressWebView.apply {
-            addJavascriptInterface(WebViewData(), "Leaf")
-            webViewClient = client
-            webChromeClient = chromeClient
-            loadUrl("http://54.180.8.0:8080/daum.html")
-        }
-    }
-
-    private val client: WebViewClient = object : WebViewClient() {
-
-        override fun shouldOverrideUrlLoading(
-            view: WebView?,
-            request: WebResourceRequest?
-        ): Boolean {
-            return false
-        }
-
-        @SuppressLint("WebViewClientOnReceivedSslError")
-        override fun onReceivedSslError(
-            view: WebView?,
-            handler: SslErrorHandler?,
-            error: SslError?
-        ) {
-            handler?.proceed()
-        }
-    }
-
-    private inner class WebViewData {
-        @SuppressLint("SetTextI18n")
-        @JavascriptInterface
-        fun getAddress(zoneCode: String, roadAddress: String, buildingName: String) {
-            CoroutineScope(Dispatchers.Default).launch {
-                withContext(CoroutineScope(Dispatchers.Main).coroutineContext) {
-                    binding.etFindAddress.setText("$roadAddress $buildingName")
-                }
-            }
-        }
-    }
-
-    private val chromeClient = object : WebChromeClient() {
-
-        @SuppressLint("SetJavaScriptEnabled")
-        override fun onCreateWindow(
-            view: WebView?,
-            isDialog: Boolean,
-            isUserGesture: Boolean,
-            resultMsg: Message?
-        ): Boolean {
-
-            val newWebView = WebView(requireContext()).apply {
-                settings.javaScriptEnabled = true
-            }
-
-            val dialog = Dialog(requireContext())
-
-            dialog.setContentView(newWebView)
-
-            val params = dialog.window!!.attributes.apply {
-                width = ViewGroup.LayoutParams.MATCH_PARENT
-                height = ViewGroup.LayoutParams.MATCH_PARENT
-            }
-
-            dialog.window!!.attributes = params
-            dialog.show()
-
-            newWebView.webChromeClient = object : WebChromeClient() {
-                override fun onJsAlert(
-                    view: WebView,
-                    url: String,
-                    message: String,
-                    result: JsResult
-                ): Boolean {
-                    super.onJsAlert(view, url, message, result)
-                    return true
-                }
-
-                override fun onCloseWindow(window: WebView?) {
-                    dialog.dismiss()
-                }
-            }
-
-            (resultMsg!!.obj as WebView.WebViewTransport).webView = newWebView
-            resultMsg.sendToTarget()
-
-            return true
-        }
-    }
-}
