--- android/app/src/main/res/layout/activity_transfer.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.transfer.TransferActivity">
-
-        <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/al_appbar_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <com.google.android.material.appbar.MaterialToolbar
-                android:id="@+id/topAppBar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                app:title="내 방 양도하기"
-                app:titleTextColor="@color/black"
-                app:navigationIcon="@drawable/ic_close"
-                app:navigationIconTint="@color/gray"
-                android:backgroundTint="@color/white"
-                style="@style/Widget.MaterialComponents.Toolbar.Primary"
-                />
-
-        </com.google.android.material.appbar.AppBarLayout>
-
-
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/fragment_container_view_in_transfer"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toBottomOf="@id/al_appbar_layout"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:name="androidx.navigation.fragment.NavHostFragment"
-            app:defaultNavHost="true"
-            app:navGraph="@navigation/transfer_graph"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
