+++ android/app/src/main/res/layout/fragment_address_search.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="com.example.home_rent_app.ui.viewmodel.TransferViewModel" />
+            type="com.example.home_rent_app.ui.viewmodel.RentHomeViewModel" />
 
     </data>
 
@@ -16,7 +16,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.transfer.step3..AddressSearchFragment">
+        tools:context=".ui.renthome.step3..AddressSearchFragment">
 
         <WebView
             android:id="@+id/kakao_address_web_view"
